a lot changes and progress

+++ server/controller/Section.js
@@ -56,19 +56,25 @@ exports.createSection = async (req, res) => {
 exports.updateSection = async (req, res) => {
   try {
     const { sectionName, sectionId, courseId } = req.body;
-    console.log(sectionName, sectionId);
     const section = await Section.findByIdAndUpdate(
       sectionId,
       { sectionName },
       { new: true }
     );
-    const updatedCourse = await Course.findById(courseId)
-      .populate({ path: "courseContent", populate: { path: "subSection" } })
+
+    const course = await Course.findById(courseId)
+      .populate({
+        path: "courseContent",
+        populate: {
+          path: "SubSection",
+        },
+      })
       .exec();
+
     res.status(200).json({
       success: true,
-      message: "Section updated successfully",
-      updatedCourse,
+      message: section,
+      data: course,
     });
   } catch (error) {
     console.error("Error updating section:", error);
@@ -83,14 +89,39 @@ exports.updateSection = async (req, res) => {
 exports.deleteSection = async (req, res) => {
   try {
     const { sectionId, courseId } = req.body;
+    await Course.findByIdAndUpdate(courseId, {
+      $pull: {
+        courseContent: sectionId,
+      },
+    });
+    const section = await Section.findById(sectionId);
+    console.log(sectionId, courseId);
+    if (!section) {
+      return res.status(404).json({
+        success: false,
+        message: "Section not Found",
+      });
+    }
+
+    //delete sub section
+    await SubSection.deleteMany({ _id: { $in: section.SubSection } });
+
     await Section.findByIdAndDelete(sectionId);
-    const updatedCourse = await Course.findById(courseId)
-      .populate({ path: "courseContent", populate: { path: "subSection" } })
+
+    //find the updated course and return
+    const course = await Course.findById(courseId)
+      .populate({
+        path: "courseContent",
+        populate: {
+          path: "SubSection",
+        },
+      })
       .exec();
+
     res.status(200).json({
       success: true,
       message: "Section deleted",
-      updatedCourse,
+      data: course,
     });
   } catch (error) {
     console.error("Error deleting section:", error);
